+++ src/dirmaster.c
@@ -1081,7 +1081,7 @@ static void afc_dirmaster_internal_size2string(DirMaster *dm, char *dest, unsign
 		if (dm->size_decimals)
 		{
 			snprintf(buf, 20, "%lu", size - (rbase * rsize));
-			if (dm->size_decimals < sizeof(buf)) buf[dm->size_decimals] = 0;
+			if ((size_t)dm->size_decimals < sizeof(buf)) buf[dm->size_decimals] = 0;
 			snprintf(dest, 20, "%d.%.*s %s", rsize, dm->size_decimals, buf, afc_dirmaster_size_bases[count]);
 		}
 		else